Remove "How to be Stylish" page from golf example.
The golf example is based off of Rustici Software's SCORM example courses.
Recently, it came to light that the final page of the course (before the quiz)
was better off removed from the example courses. The page in question was
titled "How to be Stylish", and was dedicated to women and girls learning to
play golf, with text taken from WikiHow. Unfortunately, the article seemed to
have a very condescending tone, and could have been interpreted to suggest
that women don't know anything about sports or don't know how to take sports
seriously.

Because the page could be read in such a negative light, it seemed
innappropriate to include it in something that is being used as a promotional
tool (either for Rustici Software or for xAPI evangelism more generally).